backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
requests>=2.32.3
python-whois>=0.8.0
dnspython>=2.6.1
transformers>=4.40.0
torch>=2.2.0
pytest>=8.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
tldextract>=5.1.1
scikit-learn>=1.4.0
joblib>=1.4.0
python-multipart>=0.0.9
langchain>=0.3.0
langchain-groq>=0.2.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
